Separate endpoint-related shared examples
