+++ spec/controllers/api/v1/users_controller_spec.rb
@@ -1,10 +1,10 @@
 require 'rails_helper'
-require_relative '../../../support/api/v1/index'
-require_relative '../../../support/api/v1/show'
-require_relative '../../../support/api/v1/create'
-require_relative '../../../support/api/v1/update'
-require_relative '../../../support/api/v1/destroy'
 require_relative '../../../support/api/v1/request'
+require_relative '../../../support/api/v1/endpoints/index'
+require_relative '../../../support/api/v1/endpoints/show'
+require_relative '../../../support/api/v1/endpoints/create'
+require_relative '../../../support/api/v1/endpoints/update'
+require_relative '../../../support/api/v1/endpoints/destroy'
 
 RSpec.describe Api::V1::UsersController, type: :controller do
   include_context "authenticate requests using valid token"
